fix: make openStream work on Windows 7
Even though the PortAudio's initialization and openStream were
called on the same thread, openStream would fail due
with an "Invalid Device" error.

It worked OK on Windows 10 but but not on Windows 7 machines.

print statements indicated that "CoInitialize needs to be called".

Using COINIT_MULTITHREADED instead of COINIT_APARTMENTTHREADED fixes the
issue on my end.